requirements.txt
python-telegram-bot==20.7
google==3.0.0 